+++ mods/BrewLAN_Plenae/CityGeneration/units/ZZCITYBLOCK/ZZCITYBLOCK_script.lua
@@ -1,0 +1,5 @@
+local CStructureUnit = import('/lua/cybranunits.lua').CStructureUnit
+
+ZZCITYBLOCK = Class(CStructureUnit) {}
+
+TypeClass = ZZCITYBLOCK